fixing problems with how I imported data
I used two "slightly" different ways to generate the y array in ex2 and ex2_reg. The result was one had the dimensions (100,) and the other (100,1) when this went through sigmoid and then into the cost functions, I would get arrays of very unexpected sizes, (118,118) vs (118,1), and values.  I matched the two data import methods and changed the sigmoid function and now everything is working correctly in both